backend/flask/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
sqlalchemy.orm
